basic collision is working

+++ components/SizeComponent.go
@@ -1,6 +1,6 @@
 package components
 
 type SizeComponent struct {
-	Width  int32
-	Height int32
+	Width  float32
+	Height float32
 }
